Исправил недочеты ментора и распределил по папочкам
