+++ Auctix/AppDelegate.swift
@@ -16,9 +16,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
         Thread.sleep(forTimeInterval: 1.0)
-        UITabBar.appearance().tintColor = .honeyYellow//нужный цвет
-        UITabBar.appearance().unselectedItemTintColor = .lightCornflowerBlue
-        
      
 //        UITabBarItem.appearance().setTitleTextAttributes([NSAttributedString.Key.foregroundColor: UIColor.lightCornflowerBlue], for: .normal)
