+++ Auctix/Assets.xcassets/icons_home_back/AccTabBar.imageset/Contents.json
