+++ Auctix/Assets.xcassets/icons_home_back/BidTabBar.imageset/Contents.json
