+++ Auctix/Assets.xcassets/icons_home_back/BidTabBar.imageset/Image.png
