+++ Auctix/Assets.xcassets/icons_home_back/HomeTabBar.imageset/Contents.json
