+++ Auctix/Assets.xcassets/icons_search_back/Search.imageset/Contents.json
