+++ Auctix/ExhibitionCell/ExhibitionManager.swift
